add linter and setup github actions
