+++ index.html
@@ -51,6 +51,6 @@
         </table>
     </div>
   </div>
-  <script src="assets/main.js"></script>
+  <script type="module" src="assets/main.js"></script>
 </body>
 </html>
